various code cleanups. fix for sync repos in module installer.

+++ src/ModuleManager/ModuleListObject.h
@@ -39,6 +39,6 @@ typedef enum _ModuleTaskId {
 // --------------- methods --------------
 - (NSString *)moduleName;
 - (NSString *)moduleTypeString;
-- (int)moduleStatus;
+- (NSInteger)moduleStatus;
 
 @end
